Tentative fix for bug #2961. When we have to print two terms that should be different, we compute the externalized form for both in the current state. If ever they coincide, we do it again with the implicit arguments flag set.
In other words, if those two terms appear the same, we automatically
print them with the implicit arguments flag on.

This is a bit raw, and we may try other externalization flags, but
I assume it to be (almost) sufficient for now.

TODO: check every place where we could need such a feature, and try
other flags.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@16664 85f007b7-540e-0410-9357-904b9bb8a0f7