Universe counters on slaves are in sync with master
Simple framework for remote counters.  The slaves ask
the master for a fresh value.  On the master the thread
manager answers with a bunch of fresh values (so that further
requests can be immediately satisfied).
Remote counters are guarded with a mutex on the master,
because all slave managers as well as the master thread
can access the counter at the same time.

I know the name sucks.  These counters are remote for the slaves,
and local for the master.  I'm open to suggestions...

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@16713 85f007b7-540e-0410-9357-904b9bb8a0f7