STM: add options to disable fallbacks to ease regression testing
STM falls back to local, lazy, evaluation if the slave dies badly or
if there is a marshalling error.  Both things should never occur, but
is nice to have the system recover if a bug pops up.
Nevertheless during regression testing these fallbacks should be
disabled not to hide a new bug.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@16841 85f007b7-540e-0410-9357-904b9bb8a0f7