Partial backport of Coqide: new feedback mechanism for structured content
This amounts to a new type of message called feedback and defined in
Interface to hold structured data.  Coq sends feedback messages
asynchronously (they are all fetched, like regular messages, together
with the main response to a call) and they are related to a specific
sentence by an id.

It is revision trunk@16451.