Tacinterp: fewer use of old-style goals.
There are many functions in Tacinterp which use a goal as a proxy for the pair of an environment and an evar_map. But do not build LCF tactics with them. They don't play well with the new style of tactics.

I've changed some of them to explicitely take an environment and an evar_map instead.