Revert "Rewriting the proof monad mechanism. Now it uses pure OCaml code, without"
This reverts commit bfe1141026da70d8f59cf47b5fe61ffc20a29f3c.

Conflicts:
	proofs/proofview_monad.ml

This is potentially a temporary commit until a final decision is taken on hand-written versus extracted ocaml for the tactic monad. The hand-written implementation has a bug where the + tactical would not behave properly (I tried to find why, but couldn't: the hand-written implementation looks fine, but it isn't. Beats me).

There is a conflict with Pierre-Marie's commit 4832692: Fixing backtrace registering of various tactic-related try-with blocks.