STM: fix worker crash when doing vm_compute
Kudos to Maximes for finding the culprit in no time!

Values of type 'Pre_env.key' store in the OCaml state the 'address'
of an already evaluated constant in the VM's C state.  Such values
are not sent to work processes.  The worker is going to re-evaluate the
constant, but just once, since the cache is cleared only when the env is
marshalled (via ephemerons).