Spawn: managed processes
The Spawn and Spawned modules factor the operation of spawning
a process.  Both synchronous and asynchronous channels are supported.
Both threaded and glib like main loop models are supported.  Still,
not all combinations are truly tested not equipped with a decent API:
only async + glib and sync + thread are, since these are the models we
use for coqide<->coqtop and coqtop<->worker respectively.