Remove the -fno-defer-pop cflag
According to http://caml.inria.fr/mantis/view.php?id=6346, this flag
causes ocamlc to fail on the latest version of xcode, because clang now
errors on -fno-defer-pop. According to the same issue, -fno-defer-pop
is required for computed gotos if you're using gcc 1.xx, but not gcc
3.4.0 nor 4.4.7 (nor presumably other reasonably modern versions of
gcc). I haven't actually tested this, as I don't have a mac, but it's a
relatively small change.

This closes pull request #10.

Signed-off-by: Pierre Boutillier <pierre.boutillier@ens-lyon.org>