Injection: do not fail when arguments have sort Prop.
This historic limitation of the injection tactic does not seem to have any
precise justification. In fact, the only equalities that are not projectable is
when the arguments of the constructor have sort Set or Type and the inductive
type is in Prop (due to the restriction on pattern matching).

The command "Unset Injection On Proofs" restores the old behavior.