Add [guard] tactic.
The [guard] tactic accepts simple tests (on integers) as argument, succeeds if the test passes and fails if the test fails.

Together with [numgoal] can be used to fork on the number of goals of a tactic.

The syntax is not very robust (in particular [guard n<=1] is correct but not [guard (n<=1)]). Maybe tests should be moved to the general parser to allow for more flexibility.