Info: print the name of atomic tactics.
The printing uses the printer for interpreted tactics. It only works for tactics which are defined in the new style. There are still a few atomic tactics in tacinterp which are defined in [V82] style. Namely:

exact, apply, clear, rename, reduce, change, (mutual variant of) fix, (mutual variant of) cofix

These print placeholder names which are never reparsable and not as useful.