Fixing parsing of bullets after a "...".
The lexer parses bullets only at the beginning of sentence. In
particular, the lexer recognizes sentences (this feature was
introduced for the translator and it is still used for the
beautifier). It recognized "." but not "...'. I added "..." followed
by space or eol as a terminator of sentences. I hope this is
compatible with the rest of the code dealing with end of
sentences.

Fixed also parse_to_dot which was not aware of "...".

Maybe there are similar things to do with coqide or PG?