mltop: when a plugin is loaded store its full path in the summary
This fixes the following bug related to stm: if one passes -I to coqide, then
such flag is passes to the workers; but if one uses "Add ML LoadPath" to
extend the paths in which coq looks for plugins, this extra path was
no passed to the slaves (via the command line) nor store in the system
state and hence sent to the slaves.

With this patch, when a cmxs is loaded, its full path is stored in the
summary and hence sent to the workers as one may expect.