Revert "test-suite: New names for vars but the expected invariant is preserved"
This reverts commit cc06ffe3df0ecc023ad046a085b0751ed4161cbf.

Going back to the convention of naming bound variables in hypotheses
of the goal as in 8.4.

My arguments for the revert are:

- apply ... with (id:=...) would have to be changed too, then possibly
  breaking the compatibility
- the naming became dependent of the order of variables as in

x:nat
H:forall x0, x0=0
=====
goal

but

H:forall x, x=0
x:nat
=====
goal

Accordingly, this is all a matter of convention, since the meaning of
bindings is anyway the same in both cases.