Info: print name of calls to Ltac constants ([TacCall]).
Some particular care needed to be taken to print aliases properly.

The printing of argument is just the generic printer for [genarg]. The trouble is that (apart from being incomplete), it does not know that it's printing Ltac arguments. As a consequence, the arguments are not properly quoted (e.g. if they are tactic expressions, they are not within [ltac:(...)].